pyproject.toml
[project]
name = "thesis-mas"
version = "0.1.0"
requires-python = "==3.12.*"
dependencies = [
    "pandas>=3.0.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
]
